add tabs to blog list

+++ src/app/main/blog_app/blog_components/blogList/content.blogList.js
@@ -9,7 +9,6 @@ import ChatBubbleOutlineIcon from '@material-ui/icons/ChatBubbleOutline';
 import * as blogActions from '../../store/actions';
 import { useDispatch, useSelector } from 'react-redux';
 import { createMuiTheme, ThemeProvider } from '@material-ui/core/styles';
-import BlogTags from '../blogTags';
 import { Link } from 'react-router-dom';
 import SectionHeader from '../../sectionHeader';
 const theme = createMuiTheme();
@@ -27,7 +26,11 @@ theme.typography.h4 = {
 const useStyles = makeStyles((theme) => ({
   paper: {
     padding: theme.spacing(2),
-    borderRadius: theme.spacing(1),
+    borderRadius: theme.spacing(.5),
+    marginBottom: theme.spacing(2),
+    [theme.breakpoints.down('xs')]: {
+      marginBottom: 0,
+    }
   },
   blogInfo: {
     marginLeft: 44,
@@ -48,6 +51,10 @@ const useStyles = makeStyles((theme) => ({
     paddingLeft: theme.spacing(0),
     textTransform: 'none',
   },
+  tag: {
+    display: 'inline-block',
+    margin: '8px 8px 24px 0'
+  },
 }));
 
 function Blog(props) {
@@ -60,20 +67,32 @@ function Blog(props) {
     dispatch(blogActions.likeAndUnlikeBlogPost(id));
   }
 
-  const handleDelete = () => {
-    dispatch(blogActions.deleteOneBlogPost(props.blog.id));
+  const handleDelete = (value) => {
+    if (value === 'Delete post') {
+      dispatch(blogActions.deleteOneBlogPost(props.blog.id));
+    }
   }
+
+  const blogTags = () => props.tags.map((tag, i) => {
+    return <Typography 
+            key={i}
+            variant="caption"
+            className={classes.tag}
+          >
+            {`#${tag}`}
+          </Typography>
+  });
 
   const getColor = () => !clicked ? '#4d5760' : '#F44336';
 
   return (
     <Paper className={classes.paper} variant="outlined">
       <SectionHeader
-        fullName={props.blog.employees[0].firstName}
+        fullName='Mark Jack'
         updatedAt={props.blog.updatedAt}
         id={props.blog.id}
         buttonContent={['Edit post', 'Delete post']}
-        onClick={() => handleDelete()}
+        onClick={(value) => handleDelete(value)}
       />
       <div className={classes.blogInfo}>
         <ThemeProvider theme={theme}>
@@ -82,24 +101,23 @@ function Blog(props) {
           </Link>
         </ThemeProvider>
         <div className={classes.dFlex}>
-          {props.blog.tags && <BlogTags tags={props.blog.tags} />}
+          {blogTags()}
         </div>
         <div className={classes.dFlex}>
           <Button
             style={{color: getColor()}}
             className={classes.button}
-            startIcon={clicked ? <Favorite /> : <FavoriteBorder />}
+            startIcon={!clicked && (props.blog) ? <FavoriteBorder /> : <Favorite />}
             onClick={() => handleLike(props.blog.id)}
           >
-            Like
+            {props.blog.likes !== undefined && props.blog.likes.length === 0 ? props.likes.length : 'Likes'}
           </Button>
           <Button
             style={{color: '#4d5760'}}
             className={classes.button}
             startIcon={<ChatBubbleOutlineIcon />}
-            // onClick={() => handleLike(props.blog.id)}
           >
-            Comment
+            { props.blog.comment !== undefined && props.blog.comment.length === 0 ? 'Comment' : props.blog.comment.length }
           </Button>
         </div>
       </div>
